conditionally push up vat findings if they exist

+++ stages/publish/upload-to-s3-run.sh
@@ -33,7 +33,12 @@ cp -r "${SCAN_DIRECTORY}"/* reports/
 
 cp "${BUILD_DIRECTORY}"/"${IMAGE_FILE}".tar reports/"${CI_PROJECT_NAME}"-"${IMAGE_VERSION}".tar
 cp "${PROJECT_LICENSE}" "${PROJECT_README}" reports/
-cp "${VAT_FINDINGS}" reports/
+if [ -f "${VAT_FINDINGS}" ]; then
+    cp "${VAT_FINDINGS}" reports/
+    python3 "${PIPELINE_REPO_DIR}/stages/publish/s3_upload.py" --file "${VAT_FINDINGS}" --bucket "${S3_REPORT_BUCKET}" --dest "${BASE_BUCKET_DIRECTORY}/${IMAGE_PATH}/${IMAGE_VERSION}/${REMOTE_REPORT_DIRECTORY}/${VAT_FINDINGS}"
+else
+    echo "WARNING: ${VAT_FINDINGS} does not exist, not copying into report"
+fi
 
 # Debug
 ls reports
@@ -55,4 +60,3 @@ done
 python3 "${PIPELINE_REPO_DIR}/stages/publish/s3_upload.py" --file "${PROJECT_README}" --bucket "${S3_REPORT_BUCKET}" --dest "${BASE_BUCKET_DIRECTORY}/${IMAGE_PATH}/${IMAGE_VERSION}/${REMOTE_REPORT_DIRECTORY}/${PROJECT_README}"
 python3 "${PIPELINE_REPO_DIR}/stages/publish/s3_upload.py" --file "${PROJECT_LICENSE}" --bucket "${S3_REPORT_BUCKET}" --dest "${BASE_BUCKET_DIRECTORY}/${IMAGE_PATH}/${IMAGE_VERSION}/${REMOTE_REPORT_DIRECTORY}/${PROJECT_LICENSE}"
 python3 "${PIPELINE_REPO_DIR}/stages/publish/s3_upload.py" --file "${REPORT_TAR_NAME}" --bucket "${S3_REPORT_BUCKET}" --dest "${BASE_BUCKET_DIRECTORY}/${IMAGE_PATH}/${IMAGE_VERSION}/${REMOTE_REPORT_DIRECTORY}/${REPORT_TAR_NAME}"
-python3 "${PIPELINE_REPO_DIR}/stages/publish/s3_upload.py" --file "${VAT_FINDINGS}" --bucket "${S3_REPORT_BUCKET}" --dest "${BASE_BUCKET_DIRECTORY}/${IMAGE_PATH}/${IMAGE_VERSION}/${REMOTE_REPORT_DIRECTORY}/${VAT_FINDINGS}"
